user-auth-implemented

+++ backend/controllers/blog.controller.js
@@ -5,7 +5,7 @@ const blogGet = async (req, res) => {
   // res.send(data)
 
   try {
-    const posts = await blogModel.find().sort({ date: -1 });
+    const posts = await blogModel.find({userId : req.userId}).sort({ date: -1 });
     res.status(200).json(posts);
   } catch (err) {
     res.status(500).json({ error: "Failed to fetch posts" });
@@ -18,7 +18,7 @@ const blogCreate = async (req, res) => {
 
   const { title, description, author } = req.body;
   try {
-    const newPost = new blogModel({ title, description, author });
+    const newPost = new blogModel({ title, description, author, userId:req.userId });
     await newPost.save();
     res.status(201).json(newPost);
   } catch (err) {
@@ -33,7 +33,7 @@ const blogDelete = async (req, res) => {
 
   const { id } = req.params;
   try {
-    const deletedPost = await blogModel.findByIdAndDelete(id);
+    const deletedPost = await blogModel.findByIdAndDelete({_id:id,userId:req.userId});
     if (!deletedPost) {
       return res.status(404).json({ error: "Post not found" });
     }
@@ -52,7 +52,7 @@ const blogUpdate = async (req, res) => {
   const { title, description, author } = req.body;
   try {
     const updatedPost = await blogModel.findByIdAndUpdate(
-      id,
+      {_id:id, userId: req.userId},
       { title, description, author },
       { new: true }
     );
